split constraints and constraint
