+++ tests/bofire/domain/test_concurrency_constraints.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from bofire.domain.constraints import NChooseKConstraint
+from bofire.domain.constraint import NChooseKConstraint
 from bofire.domain.domain import Domain
 from bofire.domain.features import (
     CategoricalDescriptorInput,
