+++ tests/bofire/domain/test_domain.py
@@ -7,7 +7,7 @@ import pytest
 from pandas.testing import assert_frame_equal
 from pydantic.error_wrappers import ValidationError
 
-from bofire.domain.constraints import LinearEqualityConstraint, NChooseKConstraint
+from bofire.domain.constraint import LinearEqualityConstraint, NChooseKConstraint
 from bofire.domain.domain import Domain, get_subdomain
 from bofire.domain.features import (
     CategoricalDescriptorInput,
