+++ tests/bofire/domain/test_domain_validators.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from bofire.domain.constraints import LinearEqualityConstraint
+from bofire.domain.constraint import LinearEqualityConstraint
 from bofire.domain.domain import Domain
 from bofire.domain.features import (
     CategoricalDescriptorInput,
